0x02-Session_authentication/requirements.txt
-U flask
Flask-Cors==3.0.8
Jinja2==2.11.2
requests==2.18.4
pycodestyle==2.6.0
